move building frequency tables to its own module
